Added logic to seperate file
